* Fixed numerous compiler warnings. * Bumped version (2.0.0.svn32).
git-svn-id: https://svn.linuxsampler.org/svn@3054 d2c4447f-3961-4fff-bec3-96d84be9abe5

+++ linuxsampler/trunk/src/EventListeners.h
@@ -1,6 +1,7 @@
 /***************************************************************************
  *                                                                         *
- *   Copyright (C) 2007, 2008 Grigor Iliev                                 *
+ *   Copyright (C) 2007 - 2009 Grigor Iliev                                *
+ *   Copyright (C) 2016 Christian Schoenebeck                              *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -66,7 +67,7 @@ namespace LinuxSampler {
              * Gets the number of the registered listeners.
              */
             int GetListenerCount() {
-                return vListenerList.size();
+                return (int) vListenerList.size();
             }
             
             /**
