+++ linuxsampler/trunk/src/engines/sfz/Stream.cpp
@@ -89,7 +89,7 @@ namespace LinuxSampler { namespace sfz {
         info.ChannelsPerFrame = pRgn->pSample->GetChannelCount();
         info.FrameSize        = pRgn->pSample->GetFrameSize();
         info.BytesPerSample   = pRgn->pSample->GetFrameSize() / pRgn->pSample->GetChannelCount();
-        info.TotalSampleCount = pRgn->pSample->GetTotalFrameCount();
+        info.TotalSampleCount = (int)pRgn->pSample->GetTotalFrameCount();
 
         Sample::PlaybackState playbackState;
         playbackState.position         = SampleOffset;
